Revert "Merge branch 'main' into feature/GooglePlayLoadSave"
This reverts commit d388272bf6f0a857149ce7a3319851b4ab40b6de, reversing
changes made to 42569af39e46369ad7541745dd9441cbfda9ab27.

+++ Assets/Script/BtcPanel.cs
@@ -164,7 +164,7 @@ public class BtcPanel : MonoBehaviour
     {
         try
         {
-            double btcForBuy = scriptGameSystem.currentBtc;
+            float btcForBuy = scriptGameSystem.currentBtc;
             UInt64 moneyToGet = Convert.ToUInt64(scriptGameSystem.currentBtcPrice* btcForBuy);
 
             if (btcForBuy < 0.00000001)
